Analysis-of-body-language-and-speech-in-video-main/requirements.txt
moviepy==1.0.3
whisper==1.1.10
google-generativeai==0.8.3
colorama==0.4.6
rich==13.8.1
ipython==8.29.0